+++ CHANGELOG.md
@@ -39,6 +39,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 - [#445](https://github.com/babylonlabs-io/babylon/pull/445) Reject BTC headers
 forks starting with already known header
+- [#457](https://github.com/babylonlabs-io/babylon/pull/457)  Remove staking msg
+server and update gentx to generate `MsgWrappedCreateValidator`
 
 ## v1.0.0-rc4